Allow to store items in temp directory

+++ src/l2a_item.cpp
@@ -690,43 +690,66 @@ void L2A::CheckItemDataStructure()
         used_pdf_files.push_back(new_pdf_path);
     }
 
-    // Cleanup pdf links directory.
+    if (L2A::GlobalMutable().label_strategy_ == L2A::GLOBAL::LabelStrategy::label_strategy_temp_)
     {
-        // Get all pdf items.
+        // Nothing to do here, as we clean the temp directpory each time at startup, and when a document is opened, the
+        // items will be stored in the temp directory. The only thing we do here, is to check if a "links" directory
+        // with LaTeX2AI items exists and warn the user that this will be outdated.
+
+        // Get all PDF items in the local directory that match the LaTeX2AI naming convention.
+        const ai::FilePath pdf_file_directory_local = L2A::GlobalMutable().GetPdfFileDirectoryLocal();
         ai::UnicodeString pattern = ai::UnicodeString(".*") + L2A::NAMES::pdf_item_post_fix_ + ".*\\.pdf$";
-        std::vector<ai::FilePath> pdf_item_files = L2A::UTIL::FindFilesInFolder(pdf_file_directory, pattern);
-
-        // Get all documents parallel to the current document.
-        pattern = ai::UnicodeString(".*\\.ai$");
-        std::vector<ai::FilePath> ai_document_files =
-            L2A::UTIL::FindFilesInFolder(L2A::UTIL::GetDocumentPath().GetParent(), pattern);
-
-        // Loop over each pdf label and check if it should be deleted.
-        for (const auto& pdf_path : pdf_item_files)
+        std::vector<ai::FilePath> pdf_item_files = L2A::UTIL::FindFilesInFolder(pdf_file_directory_local, pattern);
+        if (pdf_item_files.size() > 0)
+            l2a_warning(
+                "There are " + L2A::UTIL::IntegerToString(pdf_item_files.size()) +
+                " LaTeX2AI items found in the local directory: \"" + pdf_file_directory_local.GetFullPath() +
+                "\".\nHowever, you currently have the option enabled to store items in the *temporary* folder (this "
+                "is the default).\n\nThese local items will NOT be updated while this option is active.\n"
+                "To avoid confusion or outdated content, please delete the items in the local directory.\n"
+                "If you want to use the local directory instead, you can switch to 'Local links directory' in the "
+                "LaTeX2AI options.");
+    }
+    else if (L2A::GlobalMutable().label_strategy_ == L2A::GLOBAL::LabelStrategy::label_strategy_local_)
+    {
+        // Cleanup pdf links directory.
         {
-            bool delete_this_file = true;
+            // Get all pdf items.
+            ai::UnicodeString pattern = ai::UnicodeString(".*") + L2A::NAMES::pdf_item_post_fix_ + ".*\\.pdf$";
+            std::vector<ai::FilePath> pdf_item_files = L2A::UTIL::FindFilesInFolder(pdf_file_directory, pattern);
 
-            // Check if the pdf if part of this document.
-            for (const auto& used_file : used_pdf_files)
+            // Get all documents parallel to the current document.
+            pattern = ai::UnicodeString(".*\\.ai$");
+            std::vector<ai::FilePath> ai_document_files =
+                L2A::UTIL::FindFilesInFolder(L2A::UTIL::GetDocumentPath().GetParent(), pattern);
+
+            // Loop over each pdf label and check if it should be deleted.
+            for (const auto& pdf_path : pdf_item_files)
             {
-                if (pdf_path == used_file) delete_this_file = false;
+                bool delete_this_file = true;
+
+                // Check if the pdf if part of this document.
+                for (const auto& used_file : used_pdf_files)
+                {
+                    if (pdf_path == used_file) delete_this_file = false;
+                }
+
+                // Check if the pdf fits to another Illustrator document by its name.
+                const ai::UnicodeString pdf_name = pdf_path.GetFileName();
+                for (const auto& ai_file : ai_document_files)
+                {
+                    // Do not check the name of this document, as the individual items are already checked above.
+                    if (L2A::UTIL::GetDocumentPath() == ai_file) continue;
+
+                    ai::UnicodeString ai_name = ai_file.GetFileNameNoExt();
+                    if (L2A::UTIL::StartsWith(pdf_name, ai_name + L2A::NAMES::pdf_item_post_fix_, true))
+                        delete_this_file = false;
+                }
+
+                // If the file is not used by this document or shares the name with an existing Illustrator document,
+                // delete it.
+                if (delete_this_file) L2A::UTIL::RemoveFile(pdf_path);
             }
-
-            // Check if the pdf fits to another Illustrator document by its name.
-            const ai::UnicodeString pdf_name = pdf_path.GetFileName();
-            for (const auto& ai_file : ai_document_files)
-            {
-                // Do not check the name of this document, as the individual items are already checked above.
-                if (L2A::UTIL::GetDocumentPath() == ai_file) continue;
-
-                ai::UnicodeString ai_name = ai_file.GetFileNameNoExt();
-                if (L2A::UTIL::StartsWith(pdf_name, ai_name + L2A::NAMES::pdf_item_post_fix_, true))
-                    delete_this_file = false;
-            }
-
-            // If the file is not used by this document or shares the name with an existing Illustrator document, delete
-            // it.
-            if (delete_this_file) L2A::UTIL::RemoveFile(pdf_path);
         }
     }
 }
